Fix snapshot model and add unit tests

+++ snapshot.go
@@ -19,21 +19,21 @@ type Snapshot struct {
 
 // SnapshotProperties properties
 type SnapshotProperties struct {
-	Name                string `json:"name,omitempty"`
-	Description         string `json:"description,omitempty"`
-	Location            string `json:"location,omitempty"`
-	Size                int    `json:"size,omitempty"`
-	CPUHotPlug          bool   `json:"cpuHotPlug,omitempty"`
-	CPUHotUnplug        bool   `json:"cpuHotUnplug,omitempty"`
-	RAMHotPlug          bool   `json:"ramHotPlug,omitempty"`
-	RAMHotUnplug        bool   `json:"ramHotUnplug,omitempty"`
-	NicHotPlug          bool   `json:"nicHotPlug,omitempty"`
-	NicHotUnplug        bool   `json:"nicHotUnplug,omitempty"`
-	DiscVirtioHotPlug   bool   `json:"discVirtioHotPlug,omitempty"`
-	DiscVirtioHotUnplug bool   `json:"discVirtioHotUnplug,omitempty"`
-	DiscScsiHotPlug     bool   `json:"discScsiHotPlug,omitempty"`
-	DiscScsiHotUnplug   bool   `json:"discScsiHotUnplug,omitempty"`
-	LicenceType         string `json:"licenceType,omitempty"`
+	Name                string  `json:"name,omitempty"`
+	Description         string  `json:"description,omitempty"`
+	Location            string  `json:"location,omitempty"`
+	Size                float32 `json:"size,omitempty"`
+	CPUHotPlug          bool    `json:"cpuHotPlug,omitempty"`
+	CPUHotUnplug        bool    `json:"cpuHotUnplug,omitempty"`
+	RAMHotPlug          bool    `json:"ramHotPlug,omitempty"`
+	RAMHotUnplug        bool    `json:"ramHotUnplug,omitempty"`
+	NicHotPlug          bool    `json:"nicHotPlug,omitempty"`
+	NicHotUnplug        bool    `json:"nicHotUnplug,omitempty"`
+	DiscVirtioHotPlug   bool    `json:"discVirtioHotPlug,omitempty"`
+	DiscVirtioHotUnplug bool    `json:"discVirtioHotUnplug,omitempty"`
+	DiscScsiHotPlug     bool    `json:"discScsiHotPlug,omitempty"`
+	DiscScsiHotUnplug   bool    `json:"discScsiHotUnplug,omitempty"`
+	LicenceType         string  `json:"licenceType,omitempty"`
 }
 
 //Snapshots object
